actualizacion

+++ server/vacaciones/historico/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean,Sequence
+from sqlalchemy import Column, Integer, String, DateTime, Boolean,Sequence,DECIMAL
 from sqlalchemy.sql.schema import ForeignKey
 from sqlalchemy.orm import relationship
 from datetime import datetime
@@ -19,7 +19,7 @@ class V_historico(Serializable, Base):
 
     id = Column(Integer, Sequence('id'), primary_key=True)
     fkpersona = Column(Integer, ForeignKey("ASISTENCIA.cb_rrhh_persona.id"))
-    dias = Column(Integer, nullable=True)
+    dias = Column(DECIMAL, nullable=True)
     descripcion = Column(String(255), nullable=True,default="")
     operacion = Column(String(10), nullable=True)
     fecha = Column(DateTime, nullable=False, default=fecha_zona)
